add new API to check if wifi is enabled, also fix data test
After change to TestDevice#disconnectFromWifi, attempt to
disconnect while in fact wifi is disabled will end in a failure.
Unfortunately mobile data test was using this to ensure wifi was
disabled.

To correct this, mobile data test should only assert if wifi is
in the correct status or not, and new API is added to test the
wifi state accordingly.

Change-Id: Ib4fad2b0603a034916872aadd1c91923c1292dc3