expose waitForDeviceShell in ITestDevice
IDeviceStateMonitor already has an implementation, exposing it
in ITestDevice so that it can be used to wait for shell.

Use case:
Certain devices exposes normal adb mode while it's still in
recovery. This causes device to be online, but no `adb shell`
command can be executed (if it were in recovery mode, it won't
be considered "online").

Change-Id: I9a111e20e97464c36df18db5efea0b787647d006