Add IShardableListener.
Currently if a invocation is sharded, listeners only get events when
all shards are complete. This is annoying for some use cases, like
CTS that wants to report streaming test execution status to stdout.

This CL adds the ability for a listener to be copied to each shard.
Also tweaks folder build cloning to try copying via hardlinks first.
Also add a convenience method for OptionCopier.

Change-Id: I37dc16b2d531efa0e93c2b8eeec794a258d3e32d