Move starting and stopping of logcat to invocation from devicemanager.
Previously devicemanager would start and stop background logcat capture
on allocate/free of device. This doesn't really make sense from a
design POV, and also meant the --max-tmp-logcat-file option was ignored
since logcat was started before it was set.

This CL moves start and stop of logcat to the invocation, after
the device options are set.

Bug: 7364106

Change-Id: I5d43e799f6f63a0787bf958909500a2458d5fb12