Take bugreports every 10 failures
BugreportListener doesn't work well for the telephony stress tests
because 1) we want the iteration number in the bugreport and 2) the
bugreport listener takes a bugreport after every failure. This can
result in up to 1000 bugreports for 1000 iterations which is a waste of
space and time it takes to collect the bugreports.

Change-Id: I2068ddbaf9f0ffa62b54f5fc8a81b7af80c25fb5